src/backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-dotenv>=1.0.0
firebase-admin>=6.6.0
elevenlabs>=1.0.0
twilio>=9.0.0
